refactor(blorktools): world-panel and axis-indicator directories
