+++ packages/blorktools/src/asset_debugger/world-panel/world-panel.html
@@ -1,5 +1,5 @@
 <!-- World Panel Component -->
-<link rel="stylesheet" href="../../css/world-panel.css">
+<link rel="stylesheet" href="/asset_debugger/world-panel/world-panel.css">
 <div id="world-tab" class="tab-content">
     <div class="world-section">
         <div class="collapsible-header section-header">
